Refactored Code and all isFeatureEnabled tests are getting passed
