+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/models/Constants.java
@@ -1,0 +1,7 @@
+package com.optimizely.ab.fsc_app.bdd.models;
+
+public class Constants {
+    public static final String LISTENER_CALLED = "listener_called";
+    public static final String DISPATCHED_EVENTS = "dispatched_event";
+    public static final String USER_PROFILES = "user_profiles";
+}
