+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/models/requests/GetFeatureVariableDoubleRequest.java
@@ -1,4 +1,4 @@
-package com.optimizely.ab.fsc_app.bdd.support.requests;
+package com.optimizely.ab.fsc_app.bdd.models.requests;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,23 +6,27 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Map;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetFeatureVariableBooleanRequest {
+public class GetFeatureVariableDoubleRequest {
+
     @JsonProperty("feature_flag_key")
     private String featureFlagKey;
+
     @JsonProperty("variable_key")
     private String variableKey;
+
     @JsonProperty("user_id")
     private String userId;
+
     @JsonProperty("attributes")
     private Map<String, ?> attributes;
 
-    public GetFeatureVariableBooleanRequest() {
+    public GetFeatureVariableDoubleRequest() {
     }
 
-    public GetFeatureVariableBooleanRequest(String featureFlagKey,
-                                            String variableKey,
-                                            String userId,
-                                            Map<String, ?> attributes) {
+    public GetFeatureVariableDoubleRequest(String featureFlagKey,
+                                           String variableKey,
+                                           String userId,
+                                           Map<String, ?> attributes) {
         this.featureFlagKey = featureFlagKey;
         this.variableKey = variableKey;
         this.userId = userId;
@@ -60,5 +64,4 @@ public class GetFeatureVariableBooleanRequest {
     public void setAttributes(Map<String, ?> attributes) {
         this.attributes = attributes;
     }
-
 }
