+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/models/requests/GetFeatureVariableIntegerRequest.java
@@ -1,4 +1,4 @@
-package com.optimizely.ab.fsc_app.bdd.support.requests;
+package com.optimizely.ab.fsc_app.bdd.models.requests;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
