+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/models/requests/GetFeatureVariableStringRequest.java
@@ -1,4 +1,5 @@
-package com.optimizely.ab.fsc_app.bdd.support.requests;
+package com.optimizely.ab.fsc_app.bdd.models.requests;
+
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,8 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Map;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetFeatureVariableDoubleRequest {
-
+public class GetFeatureVariableStringRequest {
     @JsonProperty("feature_flag_key")
     private String featureFlagKey;
 
@@ -20,10 +20,10 @@ public class GetFeatureVariableDoubleRequest {
     @JsonProperty("attributes")
     private Map<String, ?> attributes;
 
-    public GetFeatureVariableDoubleRequest() {
+    public GetFeatureVariableStringRequest() {
     }
 
-    public GetFeatureVariableDoubleRequest(String featureFlagKey,
+    public GetFeatureVariableStringRequest(String featureFlagKey,
                                            String variableKey,
                                            String userId,
                                            Map<String, ?> attributes) {
@@ -64,4 +64,5 @@ public class GetFeatureVariableDoubleRequest {
     public void setAttributes(Map<String, ?> attributes) {
         this.attributes = attributes;
     }
+
 }
