+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/models/requests/GetVariationRequest.java
@@ -1,5 +1,4 @@
-package com.optimizely.ab.fsc_app.bdd.support.requests;
-
+package com.optimizely.ab.fsc_app.bdd.models.requests;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -7,19 +6,27 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Map;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class IsFeatureEnabledRequest {
+public class GetVariationRequest {
 
-    @JsonProperty("feature_flag_key")
-    private String featureFlagKey;
-
+    @JsonProperty("experiment_key")
+    private String experimentKey;
     @JsonProperty("user_id")
     private String userId;
 
     @JsonProperty("attributes")
     private Map<String, ?> attributes;
 
-    public String getFeatureFlagKey() {
-        return featureFlagKey;
+    public GetVariationRequest() {
+    }
+
+    public GetVariationRequest(String experimentKey, String userId, Map<String, String> attributes) {
+        this.experimentKey = experimentKey;
+        this.userId = userId;
+        this.attributes = attributes;
+    }
+
+    public String getExperimentKey() {
+        return experimentKey;
     }
 
     public String getUserId() {
@@ -30,8 +37,8 @@ public class IsFeatureEnabledRequest {
         return attributes;
     }
 
-    public void setFeatureFlagKey(String featureFlagKey) {
-        this.featureFlagKey = featureFlagKey;
+    public void setExperimentKey(String experimentKey) {
+        this.experimentKey = experimentKey;
     }
 
     public void setUserId(String userId) {
@@ -41,5 +48,4 @@ public class IsFeatureEnabledRequest {
     public void setAttributes(Map<String, ?> attributes) {
         this.attributes = attributes;
     }
-
 }
