+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/models/requests/TrackRequest.java
@@ -1,4 +1,5 @@
-package com.optimizely.ab.fsc_app.bdd.support.requests;
+package com.optimizely.ab.fsc_app.bdd.models.requests;
+
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,27 +7,22 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Map;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetVariationRequest {
+public class TrackRequest {
 
-    @JsonProperty("experiment_key")
-    private String experimentKey;
+    @JsonProperty("event_key")
+    private String eventKey;
+
     @JsonProperty("user_id")
     private String userId;
 
     @JsonProperty("attributes")
     private Map<String, ?> attributes;
 
-    public GetVariationRequest() {
-    }
+    @JsonProperty("event_tags")
+    private Map<String, ?> eventTags;
 
-    public GetVariationRequest(String experimentKey, String userId, Map<String, String> attributes) {
-        this.experimentKey = experimentKey;
-        this.userId = userId;
-        this.attributes = attributes;
-    }
-
-    public String getExperimentKey() {
-        return experimentKey;
+    public String getEventKey() {
+        return eventKey;
     }
 
     public String getUserId() {
@@ -37,8 +33,12 @@ public class GetVariationRequest {
         return attributes;
     }
 
-    public void setExperimentKey(String experimentKey) {
-        this.experimentKey = experimentKey;
+    public Map<String, ?> getEventTags() {
+        return eventTags;
+    }
+
+    public void setEventKey(String eventKey) {
+        this.eventKey = eventKey;
     }
 
     public void setUserId(String userId) {
@@ -47,5 +47,9 @@ public class GetVariationRequest {
 
     public void setAttributes(Map<String, ?> attributes) {
         this.attributes = attributes;
+    }
+
+    public void setEventTags(Map<String, ?> eventTags) {
+        this.eventTags = eventTags;
     }
 }
