+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/models/responses/MethodResponse.java
@@ -1,4 +1,4 @@
-package com.optimizely.ab.fsc_app.bdd.support.responses;
+package com.optimizely.ab.fsc_app.bdd.models.responses;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
