+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/optlyplugins/userprofileservices/NoOpService.java
@@ -1,4 +1,4 @@
-package com.optimizely.ab.fsc_app.bdd.support.userprofileservices;
+package com.optimizely.ab.fsc_app.bdd.optlyplugins.userprofileservices;
 
 import java.util.Collection;
 import java.util.Map;
