+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/support/resources/GetEnabledFeaturesResource.java
@@ -1,12 +1,9 @@
 package com.optimizely.ab.fsc_app.bdd.support.resources;
 
-import com.optimizely.ab.config.Variation;
-import com.optimizely.ab.fsc_app.bdd.OptimizelyE2EService;
-import com.optimizely.ab.fsc_app.bdd.support.requests.GetEnabledFeaturesRequest;
-import com.optimizely.ab.fsc_app.bdd.support.requests.GetVariationRequest;
-import com.optimizely.ab.fsc_app.bdd.support.responses.BaseResponse;
-import com.optimizely.ab.fsc_app.bdd.support.responses.ListenerMethodArrayResponse;
-import com.optimizely.ab.fsc_app.bdd.support.responses.ListenerMethodResponse;
+import com.optimizely.ab.fsc_app.bdd.support.OptimizelyE2EService;
+import com.optimizely.ab.fsc_app.bdd.models.requests.GetEnabledFeaturesRequest;
+import com.optimizely.ab.fsc_app.bdd.models.responses.BaseResponse;
+import com.optimizely.ab.fsc_app.bdd.models.responses.ListenerMethodArrayResponse;
 
 import java.util.List;
