+++ fsc-app/src/androidTest/java/com/optimizely/ab/fsc_app/bdd/support/resources/GetFeatureVariableDoubleResource.java
@@ -1,11 +1,9 @@
 package com.optimizely.ab.fsc_app.bdd.support.resources;
 
-import com.optimizely.ab.config.Variation;
-import com.optimizely.ab.fsc_app.bdd.OptimizelyE2EService;
-import com.optimizely.ab.fsc_app.bdd.support.requests.GetFeatureVariableDoubleRequest;
-import com.optimizely.ab.fsc_app.bdd.support.requests.GetVariationRequest;
-import com.optimizely.ab.fsc_app.bdd.support.responses.BaseResponse;
-import com.optimizely.ab.fsc_app.bdd.support.responses.ListenerMethodResponse;
+import com.optimizely.ab.fsc_app.bdd.support.OptimizelyE2EService;
+import com.optimizely.ab.fsc_app.bdd.models.requests.GetFeatureVariableDoubleRequest;
+import com.optimizely.ab.fsc_app.bdd.models.responses.BaseResponse;
+import com.optimizely.ab.fsc_app.bdd.models.responses.ListenerMethodResponse;
 
 public class GetFeatureVariableDoubleResource extends BaseResource<Double> {
